requirements.txt
yt-dlp>=2023.11.16
pyannote.audio==3.1.1
torch==2.0.1
torchaudio==2.0.2
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0,<2.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
scipy>=1.11.0
huggingface_hub>=0.16.0